move tests from transferFeeTests to TransferFeeTest and fix breaking mintTest
